PLAT-48: Restructure the application directories
